pyproject.toml
[tool.poetry]
name = "profile-viewer"
version = "0.1.0"
description = ""
authors = ["Roman-Sergeichuk <sergeichuk.roman@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.1.6"
airtable-python-wrapper = "^0.15.1"
python-dotenv = "^0.15.0"
argparse = "^1.4.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
sync = "airtable_sync:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
